Rename and move utils file to be more descriptive
